Add files via upload
I designed a website for a gym using HTML and CSS, creating a sleek and modern design that aligns with the gym's brand identity. The website features high-quality images, clear calls-to-action, and a user-friendly layout to encourage website visitors to sign up for memberships or classes. I used responsive design techniques to ensure the website looks great and functions well on desktop and mobile devices. Through the use of clean and efficient code, the website loads quickly and performs well, providing a smooth browsing experience. Overall, the website design effectively communicates the gym's message and offerings to potential customers, and I am proud to have created it.